+++ randovania/games/samus_returns/gui/preset_settings/msr_goal_tab.py
@@ -6,8 +6,12 @@ from typing import TYPE_CHECKING
 from PySide6 import QtCore
 
 from randovania.games.samus_returns.gui.generated.preset_msr_goal_ui import Ui_PresetMSRGoal
-from randovania.games.samus_returns.layout.msr_configuration import MSRArtifactConfig, MSRConfiguration
-from randovania.gui.lib import signal_handling
+from randovania.games.samus_returns.layout.msr_configuration import (
+    FinalBossConfiguration,
+    MSRArtifactConfig,
+    MSRConfiguration,
+)
+from randovania.gui.lib.signal_handling import on_checked, set_combo_with_value
 from randovania.gui.preset_settings.preset_tab import PresetTab
 
 if TYPE_CHECKING:
@@ -27,12 +31,18 @@ class PresetMSRGoal(PresetTab, Ui_PresetMSRGoal):
         self.goal_layout.setAlignment(QtCore.Qt.AlignmentFlag.AlignTop)
         self.restrict_placement_radiobutton.toggled.connect(self._on_restrict_placement)
         self.free_placement_radiobutton.toggled.connect(self._on_free_placement)
-        signal_handling.on_checked(self.prefer_metroids_check, self._on_prefer_metroids)
-        signal_handling.on_checked(self.prefer_stronger_metroids_check, self._on_prefer_stronger_metroids)
-        signal_handling.on_checked(self.prefer_bosses_check, self._on_prefer_bosses)
+
+        on_checked(self.prefer_metroids_check, self._on_prefer_metroids)
+        on_checked(self.prefer_stronger_metroids_check, self._on_prefer_stronger_metroids)
+        on_checked(self.prefer_bosses_check, self._on_prefer_bosses)
+
         self.required_slider.valueChanged.connect(self._on_required_slider_changed)
         self.placed_slider.valueChanged.connect(self._on_placed_slider_changed)
         self._update_slider_max()
+
+        for i, final_boss in enumerate(FinalBossConfiguration):
+            self.final_boss_combo.setItemData(i, final_boss)
+        self.final_boss_combo.currentIndexChanged.connect(self._on_boss_combo_changed)
 
     @classmethod
     def tab_title(cls) -> str:
@@ -135,6 +145,10 @@ class PresetMSRGoal(PresetTab, Ui_PresetMSRGoal):
 
         self._edit_config(edit)
 
+    def _on_boss_combo_changed(self, new_index: int) -> None:
+        with self._editor as editor:
+            editor.set_configuration_field("final_boss", self.final_boss_combo.currentData())
+
     def on_preset_changed(self, preset: Preset) -> None:
         assert isinstance(preset.configuration, MSRConfiguration)
         artifacts = preset.configuration.artifacts
@@ -145,3 +159,20 @@ class PresetMSRGoal(PresetTab, Ui_PresetMSRGoal):
         self.prefer_bosses_check.setChecked(artifacts.prefer_bosses)
         self.placed_slider.setValue(artifacts.placed_artifacts)
         self.required_slider.setValue(artifacts.required_artifacts)
+
+        final_boss = preset.configuration.final_boss
+        if final_boss == FinalBossConfiguration.ARACHNUS:
+            self.boss_info_label.setText("After defeating Arachnus, you must re-enter Dam Exterior to finish.")
+        elif final_boss == FinalBossConfiguration.DIGGERNAUT:
+            self.boss_info_label.setText("The Grapple block in Area 6 - Transport to Area 7 will be moved by default.")
+        elif final_boss == FinalBossConfiguration.QUEEN:
+            self.boss_info_label.setText(
+                "To fight the Queen, you must also collect Ice Beam and defeat all 10 Larva Metroids.\n"
+                "The wall behind the Queen will be removed by default."
+            )
+        elif final_boss == FinalBossConfiguration.RIDLEY:
+            self.boss_info_label.setText("You must also collect the Baby Metroid.")
+        else:
+            self.boss_info_label.setText("This option will randomly choose one of the four bosses to fight.")
+
+        set_combo_with_value(self.final_boss_combo, final_boss)