+++ randovania/layout/preset_migration.py
@@ -1203,6 +1203,13 @@ def _migrate_v89(preset: dict) -> dict:
     return preset
 
 
+def _migrate_v90(preset: dict) -> dict:
+    if preset["game"] == "samus_returns":
+        preset["configuration"]["final_boss"] = "Ridley"
+
+    return preset
+
+
 _MIGRATIONS = [
     _migrate_v1,  # v1.1.1-247-gaf9e4a69
     _migrate_v2,  # v1.2.2-71-g0fbabe91
@@ -1293,6 +1300,7 @@ _MIGRATIONS = [
     _migrate_v87,
     _migrate_v88,  # dread freesink
     _migrate_v89,  # msr configurable required dna
+    _migrate_v90,  # msr configurable final boss
 ]
 CURRENT_VERSION = migration_lib.get_version(_MIGRATIONS)